# src/main/resources/APISpecification.yaml
openapi: 3.0.3
info:
  title: API управления номенклатурой пищевых и биологически активных добавок
  description: Система управления базой данных номенклатуры пищевых добавок и БАД
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/manufacturers:
    get:
      tags:
        - Manufacturers
      summary: Получить всех производителей
      operationId: getAllManufacturers
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Manufacturer'
    post:
      tags:
        - Manufacturers
      summary: Создать нового производителя
      operationId: createManufacturer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Manufacturer'
      responses:
        '200':
          description: Производитель успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Manufacturer'
  /api/manufacturers/{id}:
    get:
      tags:
        - Manufacturers
      summary: Получить производителя по ID
      operationId: getManufacturerById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Manufacturer'
        '404':
          description: Производитель не найден
    put:
      tags:
        - Manufacturers
      summary: Обновить производителя
      operationId: updateManufacturer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Manufacturer'
      responses:
        '200':
          description: Производитель успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Manufacturer'
    delete:
      tags:
        - Manufacturers
      summary: Удалить производителя
      operationId: deleteManufacturer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Производитель успешно удален

  # ==================== Products ====================
...
components:
  schemas:
...
    Manufacturer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          description: Наименование производителя (обязательное)
        country:
          type: string
          description: Страна производства (обязательное)
        legalAddress:
          type: string
          description: Юридический адрес
        inn:
          type: string
          description: ИНН
        ogrn:
          type: string
          description: ОГРН
        contactPhone:
          type: string
          description: Контактный телефон
        contactEmail:
          type: string
          description: Email
        website:
          type: string
          description: Веб-сайт
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...